create: component file comment and reply
