+++ src/components/UI/Comment.jsx
@@ -1,0 +1,31 @@
+// Comment component
+const Comment = ({ comment }) => {
+  return (
+    <article className="px-8">
+      <div className="flex items-center justify-between flex-wrap">
+        <div className="flex gap-3 items-center md:gap-8">
+          <img
+            src={comment.user.image}
+            alt={comment.user.name}
+            className="w-10 h-10 rounded-full"
+          />
+          <div className="flex flex-col">
+            <h1 className="m-0 text-title text-[13px] font-bold">
+              {comment.user.name}
+            </h1>
+            <p className="m-0 text-paragraph text-[13px]">
+              @{comment.user.username}
+            </p>
+          </div>
+        </div>
+        <button className="text-[#4661E6] text-[13px] font-semibold">
+          Reply
+        </button>
+      </div>
+      <p className="text-paragraph text-[15px] py-[17px] pl-14">
+        {comment.content}
+      </p>
+    </article>
+  );
+};
+export default Comment;
