+++ src/pages/FeedBack.jsx
@@ -12,6 +12,8 @@ import leftArrow from "/assets/shared/icon-arrow-left.svg";
 
 // Custom component imports
 import FeedbackContainer from "../components/UI/FeedbackContainer";
+import Comment from "../components/UI/Comment";
+import Reply from "../components/UI/Reply";
 
 // Feedback Page Component
 function FeedBack() {
@@ -38,65 +40,7 @@ function FeedBack() {
     // Set feedback data to state
     setFeedback(filtered);
   }, []);
-  // Comment component
-  const Comment = ({ comment }) => {
-    return (
-      <article className="px-8">
-        <div className="flex items-center justify-between flex-wrap">
-          <div className="flex gap-3 items-center md:gap-8">
-            <img
-              src={comment.user.image}
-              alt={comment.user.name}
-              className="w-10 h-10 rounded-full"
-            />
-            <div className="flex flex-col">
-              <h1 className="m-0 text-title text-[13px] font-bold">
-                {comment.user.name}
-              </h1>
-              <p className="m-0 text-paragraph text-[13px]">
-                @{comment.user.username}
-              </p>
-            </div>
-          </div>
-          <button className="text-[#4661E6] text-[13px] font-semibold">
-            Reply
-          </button>
-        </div>
-        <p className="text-paragraph text-[15px] py-[17px]">
-          {comment.content}
-        </p>
-      </article>
-    );
-  };
 
-  // Reply component
-  const Reply = ({ reply }) => {
-    return (
-      <article className="pl-20 pr-8">
-        <div className="flex items-center justify-between flex-wrap">
-          <div className="flex gap-3 items-center md:gap-8">
-            <img
-              src={reply.user.image}
-              alt={reply.user.name}
-              className="w-10 h-10 rounded-full"
-            />
-            <div className="flex flex-col">
-              <h1 className="m-0 text-title text-[13px] font-bold">
-                {reply.user.name}
-              </h1>
-              <p className="m-0 text-paragraph text-[13px]">
-                @{reply.user.username}
-              </p>
-            </div>
-          </div>
-          <button className="text-[#4661E6] text-[13px] font-semibold">
-            Reply
-          </button>
-        </div>
-        <p className="text-paragraph text-[15px] py-[17px]">{reply.content}</p>
-      </article>
-    );
-  };
 
   return (
     <div className="bg-bodyC p-6 min-h-screen">
@@ -122,7 +66,7 @@ function FeedBack() {
         <Row>
           <main className="my-5 bg-containerBg rounded-lg p-6">
             <h4 className="text-title text-[18px] font-bold tracking-[0.25px] pl-6">
-              {feedback?.comments?.length} Comments
+              {!feedback?.comments?.length ? 'No Comments yet' : `${feedback?.comments?.length} Comments`}
             </h4>
             {feedback?.comments?.map((comment, index) => (
               <div key={comment.id}>
@@ -131,7 +75,10 @@ function FeedBack() {
                 {/* Render replies */}
                 {comment.replies &&
                   comment.replies.map((reply) => (
-                    <Reply key={reply.id} reply={reply} />
+                    <div className="flex relative h-full">
+                      <div className={`w-[0.7px] h-[150%] bg-[#647196] opacity-[.1] absolute bottom-0 left-[51px]`}/>
+                      <Reply key={reply.id} reply={reply} />
+                    </div>
                   ))}
                 {index !== feedback?.comments?.length - 1 && <hr />}
               </div>
